投稿に紐づくタグが存在する場合は削除できないように修正

+++ app/Http/Controllers/TagsController.php
@@ -91,6 +91,13 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0) {
+
+            session()->flash('error', 'Tag can not be deleted, because it is associated to some posts.');
+
+            return redirect()->back();
+        }
+
         $tag->delete();
 
         session()->flash('success', 'Tag deleted Successfully');
